+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 - `clone` function to `Slice`
+- `asSlice` to `Builder` and `Cell`
+- `asCell` to `Builder` and `Slice`
+- `asBuilder` to `Cell` and `Slice`
 
 ## [0.6.0] - 2022-12-27